feat: add unread notifications indicator
