+++ libs/shared-components/src/mini-overlay/assets/icons/BellWithCircle.tsx
@@ -1,21 +1,21 @@
 import React from 'react'
 
-const BellWithCircle = () => (
+const BellWithCircle = ({ isPrimary }: { isPrimary: boolean }) => (
   <svg xmlns="http://www.w3.org/2000/svg" width="15" height="16" viewBox="0 0 15 16" fill="none">
     <path
       d="M9.4375 7.81817C9.4375 6.95019 9.08982 6.11776 8.47095 5.504C7.85208 4.89024 7.01271 4.54544 6.1375 4.54544C5.26229 4.54544 4.42292 4.89024 3.80405 5.504C3.18518 6.11776 2.8375 6.95019 2.8375 7.81817C2.8375 11.6364 1.1875 12.7273 1.1875 12.7273H11.0875C11.0875 12.7273 9.4375 11.6364 9.4375 7.81817Z"
-      fill="#7A818B"
-      stroke="#7A818B"
+      fill={isPrimary ? 'white' : '#7A818B'}
+      stroke={isPrimary ? 'white' : '#7A818B'}
       strokeLinecap="round"
       strokeLinejoin="round"
     />
     <path
       d="M7.08855 14.9091C6.99185 15.0744 6.85306 15.2116 6.68607 15.307C6.51908 15.4024 6.32976 15.4526 6.13705 15.4526C5.94434 15.4526 5.75501 15.4024 5.58802 15.307C5.42103 15.2116 5.28224 15.0744 5.18555 14.9091"
-      stroke="#7A818B"
+      stroke={isPrimary ? 'white' : '#7A818B'}
       strokeLinecap="round"
       strokeLinejoin="round"
     />
-    <circle cx="11.375" cy="3.5" r="3" fill="#D9304F" stroke="#F8F9FF" />
+    <circle cx="11.375" cy="3.5" r="3" fill={isPrimary ? 'white' : '#D9304F'} stroke={'white'} />
   </svg>
 )
