+++ libs/shared-components/src/mini-overlay/assets/icons/index.ts
@@ -4,5 +4,6 @@ import StopTopIcon from './StopTop'
 import PlayCenterIcon from './PlayCenter'
 import StopCenterIcon from './StopCenter'
 import BellIcon from './Bell'
+import BellWithCircle from './BellWithCircle'
 
-export { MutationFallbackIcon, ArrowIcon, StopTopIcon, PlayCenterIcon, StopCenterIcon, BellIcon }
+export { MutationFallbackIcon, ArrowIcon, StopTopIcon, PlayCenterIcon, StopCenterIcon, BellIcon, BellWithCircle }
